Keep GD&T frame defaults test lightweight
