+++ src/machines/modelingSharedTypes.ts
@@ -62,6 +62,16 @@ export type Selections = {
   graphSelections: Array<Selection>
 }
 
+export function isEnginePrimitiveSelection(
+  selection: Selections['otherSelections'][number]
+): selection is EnginePrimitiveSelection {
+  return (
+    typeof selection === 'object' &&
+    'type' in selection &&
+    selection.type === 'enginePrimitive'
+  )
+}
+
 export type SetSelections =
   | {
       selectionType: 'singleCodeCursor'
